Cleared the code and enhanced the performance of a few listeners
